[Refactor] comment 리펙토링
